requirements.txt
torch
torchvision
fastapi
uvicorn
pillow
numpy
scikit-learn
scikit-image
matplotlib
seaborn
tqdm
lime
jupyter
ipykernel
pandas
openmeteo-requests
requests-cache
retry-requests
statsmodels
